fix(tests): isolate threshold sidecar policy
Source quality collects the static architecture suite without installing the production package. Keep the shared threshold predicate in a dependency-free test module so that contract remains independently runnable.
